pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for the RuneBook project"
authors = [
    { name = "Yanis Bouchilloux", email = "bouchilloux.yanis@gmail.com" }
]
license = { text = "MIT" }
readme = "../README.md"
requires-python = ">=3.10,<4.0"

# Dépendances principales
dependencies = [
    "flask>=2.3.0",
    "python-dotenv>=1.0.0",  # pratique pour gérer les variables d'environnement
    "jwt (>=1.4.0,<2.0.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)"
]

# Dépendances de dev
[project.optional-dependencies]
dev = [
    "pytest>=7.0",       # pour les tests
    "black>=23.0",       # pour formater le code
    "flake8>=6.0"        # pour linting
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
